refactor validate phases
